+++ gantry/constants.py
@@ -8,10 +8,10 @@ CONDA_FILE = "environment.yml"
 
 CONDA_FILE_ALTERNATE = "environment.yaml"
 
-PIP_REQUIREMENTS_FILE = "requirements.txt"
-
 RUNTIME_DIR = "/gantry-runtime"
 
 RESULTS_DIR = "/results"
+
+RUNTIME_DIR = "/gantry-runtime"
 
 METRICS_FILE = f"{RESULTS_DIR}/metrics.json"